+++ oauth2/tokengenerator.py
@@ -12,6 +12,7 @@ class TokenGenerator(object):
     """
     Base class of every token generator.
     """
+
     def __init__(self):
         """
         Create a new instance of a token generator.
@@ -19,19 +20,22 @@ class TokenGenerator(object):
         self.expires_in = {}
         self.refresh_expires_in = 0
 
-    def create_access_token_data(self, grant_type):
+    def create_access_token_data(self, data, grant_type, user_id):
         """
         Create data needed by an access token.
 
         :param grant_type:
         :type grant_type: str
 
-        :return: A ``dict`` containing he ``access_token`` and the
+        :return: A ``dict`` containing the ``access_token`` and the
                  ``token_type``. If the value of ``TokenGenerator.expires_in``
                  is larger than 0, a ``refresh_token`` will be generated too.
         :rtype: dict
+
+        .. versionchanged:: 1.1.0
+            New parameters ``data`` and ``user_id``
         """
-        result = {"access_token": self.generate(), "token_type": "Bearer"}
+        result = {"access_token": self.generate(data, user_id), "token_type": "Bearer"}
 
         if self.expires_in.get(grant_type, 0) > 0:
             result["refresh_token"] = self.generate()
@@ -40,11 +44,14 @@ class TokenGenerator(object):
 
         return result
 
-    def generate(self):
+    def generate(self, data=None, user_id=None):
         """
         Implemented by generators extending this base class.
 
         :raises NotImplementedError:
+
+        .. versionchanged:: 1.1.0
+            New parameters ``data`` and ``user_id``
         """
         raise NotImplementedError
 
@@ -53,11 +60,12 @@ class URandomTokenGenerator(TokenGenerator):
     """
     Create a token using ``os.urandom()``.
     """
+
     def __init__(self, length=40):
         self.token_length = length
         TokenGenerator.__init__(self)
 
-    def generate(self):
+    def generate(self, data=None, user_id=None):
         """
         :return: A new token
         :rtype: str
@@ -74,7 +82,8 @@ class Uuid4(TokenGenerator):
     """
     Generate a token using uuid4.
     """
-    def generate(self):
+
+    def generate(self, data=None, user_id=None):
         """
         :return: A new token
         :rtype: str